Fix issue with p:THIS replacement in object types
We were adding every visited type node to the visited list to prevent
circular dependencies. However, this was causing the transform to stop
when it hit the 2nd p:THIS since all p:THIS are the same instance.

Turns out, we don't need to add every visited type to the list to
prevent circular dependencies. Since otypes are the only ones that can
cause a circular dependency (at the moment) we only need to add them to
the list.